backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
PyJWT==2.8.0
